+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added some named query errors for check results from query service client
+
 ## v3.125.1
 * Renamed `ydb_go_sdk_ydb_table_pool_node_hint_miss` and `ydb_go_sdk_ydb_query_pool_node_hint_miss` metrics to 
 `ydb_go_sdk_ydb_table_pool_node_hint` and `ydb_go_sdk_ydb_query_pool_node_hint`. Added `hit` label for them to